requirements.txt
absl-py==0.7.1
astor==0.8.0
atari-py==0.2.6
autopep8==1.4.4
backcall==0.1.0
certifi==2019.6.16
Click==7.0
cloudpickle==1.2.1
decorator==4.4.0
entrypoints==0.3
flake8==3.7.8
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gym==0.14.0
h5py==2.9.0
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mccabe==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
torchcontrib==0.0.2
torchvision==0.4.0
tqdm==4.33.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
